pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-cors = "^5.0.0"
flask = "^3.0.3"
whisper = "^1.1.10"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.151.0"
apscheduler = "^3.10.4"
python-ffmpeg = "^2.0.12"
setuptools-rust = "^1.10.2"
tiktoken = "^0.8.0"
flask-sqlalchemy = "^3.1.1"
numba = "0.59.1"
openai-whisper = "^20240930"
flask-migrate = "^4.0.7"
python-dateutil = "^2.9.0.post0"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
